pyproject.toml
[project]
name = "tp2-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "httpx>=0.28.1",
    "litellm>=1.80.10",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
